+++ examples/5qubit_gates.jl
@@ -19,9 +19,7 @@ function decompose_RX_on_5qubits()
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",
     
-    "optimizer" => "gurobi",
-    "optimizer_presolve" => true, #turning this true will give infeasiblity in cplex - most probably a bug in cplex's presolve
-    
+    "optimizer" => "gurobi"
     )
 
     return params